Léxico/requirements.txt
selenium==4.9.1
chromedriver-autoinstaller==0.4.0
python-dotenv==1.0.0
transformers==4.29.2
tensorflow==2.12.0
SQLAlchemy==2.0.4
SQLAlchemy-Utils==0.40.0
pandas==1.5.3
PyMySQL==1.0.2
cryptography==40.0.2